Implemented complex item rotation
The rotation logic works, but there is a bug in which, when an item of a certain shape is rotated, all of the items of that shape in the inventory are rotated too. This might be due to the way the logic with shapes work and a new 2D array might be needed.