Cloud Saving and Loading using Playfab now works!
It is implemented. The problem was the async nature of Playfab while loading. Had to add an async method in the interface which allowed the script to run operations async. The only problem is that now there are two empty methods in the LocalDataManaging system and the cloud one. But I'm happy and that's all that matters.